refactor tokio pending_connection
